requirements.txt
langchain
langgraph
langchain-google-genai
faiss-cpu
streamlit
python-dotenv
requests
google-generativeai
